pyproject.toml
[tool.poetry]
name = "sscs"
version = "0.1.0"
description = ""
authors = ["Dylan Wilson <dylanjw@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.2"
django-tailwind = "^2.0.1"
django-crispy-forms = "^1.11.2"
django-bootstrap4 = "^3.0.1"
django-bootstrap-datepicker-plus = "^3.0.5"
django-jquery = "^3.1.0"
django-sockpuppet = "^0.6.0"
lxml = "^4.6.3"

[tool.poetry.dev-dependencies]
isort = "^5.8.0"

[tool.isort]
multi_line_output = 3

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
